pyproject.toml
[project]
name = "hormuz-ds"
version = "0.1.0"
description = "Hormuz Strait crisis investment decision support system"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "jinja2>=3.1",
    "matplotlib>=3.8",
    "yfinance>=0.2",
    "httpx>=0.27",
    "click>=8.1",
    "pyyaml>=6.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[project.scripts]
hormuz = "hormuz.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hormuz"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
